+++ components/portal/widgets/directives.js
@@ -27,14 +27,16 @@ define(['angular', 'require'], function(angular, require) {
   .directive('widget', ['miscService', function(miscService) {
     return {
       restrict: 'E',
-      transclude: true,
+      transclude: {
+        'removeButton': '?removeButton',
+      },
       scope: {
         fname: '@',
       },
       templateUrl: require.toUrl('./partials/widget-card.html'),
       controller: 'WidgetCardController',
       link: {
-        post: function(scope, element, attrs) {
+        post: function(scope, element, attrs, $transclude) {
           element.on('click', function(event) {
             var el = event.target;
             while (el && (!el.tagName || el.tagName.toLowerCase() !== 'a')) {
@@ -55,7 +57,9 @@ define(['angular', 'require'], function(angular, require) {
   .directive('compactWidget', ['miscService', function(miscService) {
     return {
       restrict: 'E',
-      transclude: true,
+      transclude: {
+        'removeButton': '?removeButton',
+      },
       scope: {
         fname: '@',
       },