+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 + Fixed display bug in compact widgets. Compact widgets now correctly use the per-widget contextual menu. (#794)
++ Make widget removal button focusable by keyboard when present (#797)
 
 
 ### Removed